probe: extract the shared template-loading harness
The browser and `fused` stubs, and the `new Function` trick that reaches a
function the template never exported, are needed by every probe of this
template — and a second probe is about to arrive. Moved to
template-harness.mjs rather than copied, since a duplicate would drift the
moment the template grows a new global.

No behaviour change: live-preview-probe.mjs produces the same output, and
tests/test_markdown_live_preview.py passes unchanged (24 passed).
