+++ scripts/vendor-codemirror/live-preview-probe.mjs
@@ -32,122 +32,20 @@
 //
 // Must run from this directory — module resolution needs ./node_modules.
 import fs from "node:fs";
-import {
-  EditorState, RangeSetBuilder, EditorSelection, StateEffect, StateField, Prec,
-} from "@codemirror/state";
-import { syntaxTree, ensureSyntaxTree } from "@codemirror/language";
-import {
-  Decoration, WidgetType, ViewPlugin, MatchDecorator, keymap, EditorView,
-} from "@codemirror/view";
-import { markdown, markdownLanguage, markdownKeymap } from "@codemirror/lang-markdown";
-import { autocompletion } from "@codemirror/autocomplete";
-import { indentMore, indentLess } from "@codemirror/commands";
+import { EditorState, EditorSelection } from "@codemirror/state";
+import { ensureSyntaxTree } from "@codemirror/language";
+import { markdown, markdownLanguage } from "@codemirror/lang-markdown";
+// Browser and `fused` stubs, and the trick that reaches a function the template
+// never exported. Shared with renumber-probe.mjs.
+import { loadTemplateScript } from "./template-harness.mjs";
 
-// The template runs top-level DOM and runtime code before it defines anything.
-// These stubs exist only to let it finish loading; nothing under test touches
-// them, and `_file` comes back empty so load() bails immediately.
-const stub = () => new Proxy({}, {
-  get(_target, key) {
-    if (key === "classList") return { toggle() {}, add() {}, remove() {} };
-    if (key === "dataset" || key === "style") return {};
-    if (key === "textContent" || key === "innerHTML" || key === "value") return "";
-    return () => {};
-  },
-  set() { return true; },
-});
-
-globalThis.CM = {
-  EditorState, EditorView, Decoration, WidgetType, ViewPlugin, MatchDecorator,
-  keymap, EditorSelection, RangeSetBuilder, Prec, StateEffect, StateField,
-  // A parse that has caught up with the document, which is the state the real
-  // editor converges to: `syntaxTree(state)` alone returns only what the
-  // language extension has parsed so far (a few KB on a fresh state, and it
-  // advances in the background from a view it does not have here), so a table
-  // halfway down a real note would simply not be in the tree.
-  syntaxTree: (state) => ensureSyntaxTree(state, state.doc.length, 10000)
-    || syntaxTree(state),
-  autocompletion, indentMore, indentLess, markdown, markdownLanguage,
-  markdownKeymap, basicSetup: [], oneDark: [],
-};
-// Enough of an element for a widget builder to fill in and for us to read back.
-// Not a DOM: no layout, no events, no tree semantics beyond appendChild.
-function fakeElement(tag) {
-  return {
-    tagName: tag.toUpperCase(),
-    className: "", title: "", textContent: "", href: "", src: "", alt: "",
-    type: "", checked: false, disabled: false,
-    dataset: {}, style: {}, children: [],
-    classList: { add() {}, remove() {}, toggle() {} },
-    appendChild(child) { this.children.push(child); return child; },
-    addEventListener() {},
-  };
-}
-
-globalThis.document = {
-  getElementById: stub, createElement: fakeElement, addEventListener() {},
-  querySelectorAll: () => [], documentElement: { getAttribute: () => "dark" },
-  visibilityState: "visible",
-};
-globalThis.window = { addEventListener() {}, top: { location: { pathname: "/view/x" } } };
-globalThis.MutationObserver = class { observe() {} disconnect() {} };
-globalThis.sessionStorage = { getItem: () => null, setItem() {} };
-globalThis.fusedRoBadge = { update() {} };
-globalThis.fusedGraph = { create: () => ({ setData() {}, nudge() {} }) };
 const [templatePath, docPath, caretArg, optsArg] = process.argv.slice(2);
 const opts = JSON.parse(optsArg || "{}");
 const doc = fs.readFileSync(docPath, "utf8");
 
-// A stand-in for graph.py's `note` answer, so the SCANNED state can be probed
-// too — the unscanned one is the default because that is what a mount gives.
-// The rule is deliberately crude and local to this harness: every `[[target]]`
-// in the document resolves, except one whose name says it should not. Real
-// resolution is graph.py's and is tested against graph.py (MD-6).
-const UNRESOLVED_HERE = /ghost|missing|gone/i;
-function fakeNoteScan() {
-  const links = [];
-  for (const match of doc.matchAll(/!?\[\[([^\[\]\n]+?)\]\]/g)) {
-    const target = match[1].split("|")[0].split("#")[0].trim();
-    // `[[#Heading]]` has no target: graph.py's `_resolved_links` skips it
-    // outright (it is an anchor inside this same note, not an edge), so the
-    // answer this stands in for carries no row for it either.
-    if (!target) continue;
-    const named = /\.[a-z0-9]+$/i.test(target);
-    links.push({
-      target,
-      path: UNRESOLVED_HERE.test(target)
-        ? null : "/vault/" + target + (named ? "" : ".md"),
-      title: "Title of " + target,
-      heading: "", label: "", embed: match[0].startsWith("!"), wiki: true,
-    });
-  }
-  return {
-    error: null, root: "/vault", rel: "note.md", title: "note",
-    headings: [], tags: [], links, backlinks: [],
-    notes: 1, truncated: false, skipped_large: [], parser_version: 0,
-  };
-}
-
-globalThis.fused = {
-  autoReload() {},
-  rawUrl: (p) => "/api/fs/raw?path=" + p,
-  params: {
-    get: (name) => (opts.params || {})[name] || "",
-    getAll: () => opts.params || {},
-    set() {}, onChange() {},
-  },
-  async stat() { return { mtime: 1, size: 1, writable: true }; },
-  async readFile() { return ""; },
-  async writeFile() { return { mtime: 2 }; },
-  async runPython(_py, args) {
-    if (!opts.scanned) return { error: "no_scan", message: "not scanned here" };
-    return args.action === "note" ? fakeNoteScan() : { error: "no_scan" };
-  },
-};
-
-const html = fs.readFileSync(templatePath, "utf8");
-const script = html.split("<script>\n")[1].split("</script>")[0];
-const { buildDecorations, refreshLinks } = new Function(
-  script + "\nreturn { buildDecorations, refreshLinks };")();
+const { buildDecorations, refreshLinks } = loadTemplateScript(templatePath, {
+  doc, scanned: opts.scanned === true, params: opts.params || {},
+});
 
 // The template's own load() bailed (no `_file`), so the link facts are asked for
 // explicitly here — through the real refreshLinks, so the page's own
